refactor(queries): renaming keeper file to queryKeeper
