+++ modules/assets/internal/queries/asset/queryKeeper.go
@@ -1,13 +1,14 @@
 // Copyright [2021] - [2022], AssetMantle Pte. Ltd. and the code contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package identity
+package asset
 
 import (
 	"context"
+
 	sdkTypes "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/AssetMantle/modules/modules/identities/internal/key"
+	"github.com/AssetMantle/modules/modules/assets/internal/key"
 	"github.com/AssetMantle/modules/schema/helpers"
 )
 
@@ -15,15 +16,15 @@ type queryKeeper struct {
 	mapper helpers.Mapper
 }
 
-func (queryKeeper queryKeeper) Identity(ctx context.Context, request *QueryRequest) (*QueryResponse, error) {
-	//TODO implement me
+func (queryKeeper queryKeeper) Asset(ctx context.Context, request *QueryRequest) (*QueryResponse, error) {
+	// TODO implement me
 	panic("implement me")
 }
 
 var _ helpers.QueryKeeper = (*queryKeeper)(nil)
 
 func (queryKeeper queryKeeper) Enquire(context sdkTypes.Context, queryRequest helpers.QueryRequest) helpers.QueryResponse {
-	return newQueryResponse(queryKeeper.mapper.NewCollection(context).Fetch(key.NewKey(queryRequestFromInterface(queryRequest).IdentityID)), nil)
+	return newQueryResponse(queryKeeper.mapper.NewCollection(context).Fetch(key.NewKey(queryRequestFromInterface(queryRequest).AssetID)), nil)
 }
 
 func (queryKeeper queryKeeper) Initialize(mapper helpers.Mapper, _ helpers.Parameters, _ []interface{}) helpers.Keeper {
