+++ modules/assets/internal/queries/asset/queryKeeper_test.go
