+++ modules/classifications/internal/queries/classification/queryKeeper.go
@@ -1,13 +1,14 @@
 // Copyright [2021] - [2022], AssetMantle Pte. Ltd. and the code contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package ownable
+package classification
 
 import (
 	"context"
+
 	sdkTypes "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/AssetMantle/modules/modules/splits/internal/utilities"
+	"github.com/AssetMantle/modules/modules/classifications/internal/key"
 	"github.com/AssetMantle/modules/schema/helpers"
 )
 
@@ -15,15 +16,15 @@ type queryKeeper struct {
 	mapper helpers.Mapper
 }
 
-func (queryKeeper queryKeeper) Ownable(ctx context.Context, request *QueryRequest) (*QueryResponse, error) {
-	//TODO implement me
+func (queryKeeper queryKeeper) Classification(ctx context.Context, request *QueryRequest) (*QueryResponse, error) {
+	// TODO implement me
 	panic("implement me")
 }
 
 var _ helpers.QueryKeeper = (*queryKeeper)(nil)
 
 func (queryKeeper queryKeeper) Enquire(context sdkTypes.Context, queryRequest helpers.QueryRequest) helpers.QueryResponse {
-	return newQueryResponse(utilities.GetOwnableTotalSplitsValue(queryKeeper.mapper.NewCollection(context), queryRequestFromInterface(queryRequest).OwnableID), nil)
+	return newQueryResponse(queryKeeper.mapper.NewCollection(context).Fetch(key.NewKey(queryRequestFromInterface(queryRequest).ClassificationID)), nil)
 }
 
 func (queryKeeper queryKeeper) Initialize(mapper helpers.Mapper, _ helpers.Parameters, _ []interface{}) helpers.Keeper {
